Fix hpattern case. Use only one file function
"Solve" the h-pattern by checking a nil before calling the keepcase
function. Use multiple dispatch for the keepcase function for each case
(when it's a char and when it's an empty string).

Use a single function to read the file and call the transliterate
function, to keep the logic simpler.